+++ R/install_sys_deps.R
@@ -135,6 +135,23 @@ getRefgenieGenomes = function(genomes) {
   invisible()}
 
 
+getSysDeps = function(outputDir, params) {
+  commandsDt = checkDefaultCommands(TRUE)
+
+  for (i in seq_len(nrow(commandsDt))) {
+    cmdName = gsub('_|-', '', commandsDt[i]$command)
+    paramCmdVal = params[[cmdName]]$cmd
+    if (cmdName %in% c('prefetch', 'fasterqdump', 'pigz')) {
+      paramCmdVal = params$fetch[[glue('{cmdName}Cmd')]]}
+    if (!is.null(paramCmdVal)) {
+      ver = getCommandVersion(paramCmdVal, commandsDt[i]$idx)
+      commandsDt[i, `:=`(path = paramCmdVal, version = ver)]}}
+
+  set(commandsDt, j = 'idx', value = NULL)
+  fwrite(commandsDt, file.path(outputDir, 'system_dependencies.tsv'), sep = '\t')
+  invisible(commandsDt)}
+
+
 #' Install seeker's system dependencies
 #'
 #' This function installs and configures the various programs required for