+++ tests/testthat/test-array-utils.R
@@ -38,6 +38,10 @@ test_that('getNaiveEsetGeo', {
     dir.create(outputDir)
 
     resObs = getNaiveEsetGeo(study, outputDir, file.path(outputDir, 'raw'))
+    # reduce disk size of expression set while preserving dimensions
+    Biobase::exprs(resObs$eset)[-seq(1, nrow(resObs$eset), 1000), ] = 1
+    rownames(resObs$eset@featureData@data) = NULL
+
     filesObs = dir(outputDir, recursive = TRUE)
 
     path = file.path(dataDir, paste0(study, '_eset.qs'))
@@ -61,6 +65,10 @@ test_that('getNaiveEsetAe supported', {
   dir.create(outputDir)
 
   resObs = getNaiveEsetAe(study, outputDir, file.path(outputDir, 'raw'))
+  # reduce disk size of expression set while preserving dimensions
+  Biobase::exprs(resObs$eset)[-seq(1, nrow(resObs$eset), 1000), ] = 1
+  rownames(resObs$eset@featureData@data) = NULL
+
   filesObs = dir(outputDir, recursive = TRUE)
 
   path = file.path(dataDir, paste0(study, '_eset.qs'))