+++ NEWS.md
@@ -1,3 +1,7 @@
+# seeker 1.0.0
+* Updated `seeker()` to output info of SRA Toolkit-based dependencies.
+* Revised tests and updated version number to prepare for a CRAN submission.
+
 # seeker 0.0.0.9068
 * Updated `run_seeker.sh` to use correct conda environment.
 * Updated Reproducibility vignette to not require local installation of the R package.